Fixed bug - never checked for logstash config
